Fall back to lower LOD levels in assembleImage
When the imagery and map tiles are using a different profile to the tile key, they'll need reprojecting, and if any image is unavailable, it can affect several tiles.
That also means there can be tiles which are partially covered by tiles at a particular LOD level.

Without this change, only one LOD level is tried, and if a tile is partially covered, it'll get used, but leave gaps in the map.
With this change, those gaps will be filled with lower-quality imagery from other LOD levels.