Fix static initialisation order
Previously, it was implementation-defined which static objects were initialised in which order.
In practice, that meant they were done alphabetically by translation unit when built with MSVC, which meant object factories could only be registered from files that come after Instance.cpp.

This is avoidable if the object factory map is instead a function-local static variable, as that will be initialised the first time the function's called, so can be done on demand when it's needed.