Implement system to reduce duplicate tile requests for assembleImage
Typically, images generated with assembleImage will depend on nearly the same data as their neighbours, so without this commit, many tiles will be requested over and over again.
If there's already an image on screen that depends on a certain piece of source data, it's likely that any others that are about to be generated will also need it, so if we make the results of assembleImage keep a shared pointer to their dependencies and make any still-alive images reusable, it should massively reduce network requests.

This is more effective than a basic LRU cache as it grows to accommodate all the tiles we're likely to need again and shrinks when that number shrinks instead of being a fixed size that's sometimes too small and sometimes too big.

This also eliminates the issue where an intermittent connection issue or API rate limit will lead to a fully-covered tile being replaced by one with missing data as the data we used earlier will still be available.

The other approach that was suggested (grabbing data from existing visible tiles) would give lower-quality results as resampling artefacts would build up.
It would also have been more complicated, and this is a perfectly adequate temporary solution to the problem I was experiencing.