Fix engine app startup
The engine app currently adds two TMS layers and then immediately
checks to make sure that their status isn't failed, but this causes
the app to error out and not start up correctly.  Comparing the engine
code to the demo app, it appears that the TMS layers are supposed to
be loaded later in the startup process and so they're going to start
out with the "layer closed" status.  Removing this check allows the
engine app to start up and run correctly.