Fixed several build errors which would occur when `ROCKY_HAS_IMGUI` was _not_ defined.
The `README.md` file for Rocky claims that Dear ImGui is an
"optional" requirement. Yet, upon attempting to build Rocky without
it, several build failures were encountered. The intent is for the
`ROCKY_HAS_IMGUI` macro to be used in order to determine at compile
time whether Dear ImGui-specific functionality should be included
in Rocky. However, it appears as though this macro was not being
used correctly in all places.

This commit makes some necessary adjustments in order to get Rocky
to _build_ without Dear ImGui. It should be noted that I have not
actually tested these changes outside of ensuring that the build
succeeds on my system; it may be the case that several other issues
exist which have yet to be diagnosed.