Fix washed-out colours of mosaicked image data (#103)
When converting sRGB data to floating point, it needs decoding, otherwise it produces different colours.

Adding the correct conversions to Image.cpp makes things more-or-less the right colour, but as the mosaic images were always UNORM8, dark tones would be crushed out and there'd be banding as packing colours into eight bits without sRGB encoding wastes most values on similar-looking bright colours.

To address that, the mosaic is now created with the same format as (one tile of) its source data.
Because of the fixed conversions, it's not a big deal if the source data has mixed formats.